+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* Logging of args for calls to `Logger.error()`
+* (Demo App) Use branded mastodon favicon
 
 ### v1.1.52
 * (Demo App) Log calls to `extractText()`